Updated Base Types for cross-platform support

+++ includes/base/BaseTypes.h
@@ -8,17 +8,17 @@
  *
  */
 
-typedef  signed   char  i8;
-typedef  signed   short i16;
-typedef  signed   int   i32;
-typedef  signed   long  i64;
+typedef signed char        i8;
+typedef signed short       i16;
+typedef signed int         i32;
+typedef signed long long   i64;
 
-typedef  unsigned char  ui8;
-typedef  unsigned short ui16;
-typedef  unsigned int   ui32;
-typedef  unsigned long  ui64;
+typedef unsigned char      ui8;
+typedef unsigned short     ui16;
+typedef unsigned int       ui32;
+typedef unsigned long long ui64;
 
-typedef  float          f32;
-typedef  double         f64;
+typedef  float             f32;
+typedef  double            f64;
 
 #endif // BASE_TYPES_H
